refactor(cli): extract the clap argument tree into stella-cli/src/cli.rs
main.rs sat at 1488 of the 1500-line ratchet (#629), leaving no room to
improve the help text it defines. Pure move: Cli, GlobalArgs, Command, the
auxiliary subcommand enums, and parse_budget, re-exported at the crate root
so every `crate::AuthCmd`-style path in the per-command modules is unchanged.

main.rs 1488 -> 808, cli.rs 701.
