+++ stella-cli/src/cli.rs
@@ -1,0 +1,701 @@
+//! The command-line surface: the argument tree `clap` parses, and nothing
+//! else. Extracted from `main.rs` (#629's file-size ratchet) so the shape of
+//! the CLI is one readable file, and `main.rs` is dispatch.
+//!
+//! Two audiences read this file's doc comments. The FIRST line of each
+//! variant's `///` is the summary users see in `stella --help` and in shell
+//! completion menus, so it must stand alone and stay short. Everything after
+//! the first blank `///` line is `long_about`: the full explanation, shown
+//! only by `stella help <command>`. Writing a paragraph with no blank line
+//! makes the whole paragraph the summary and buries the command list — the
+//! bug `help.rs` and `about_lines_are_short` exist to prevent.
+
+use clap::{Parser, Subcommand};
+
+use crate::{
+    build_info, commands_cmd, context_cmd, contextgraph, ingest_cmd, inspect, memory_cmd,
+    proposals_cmd, scripts_cmd, stats, storage_cmd, tune_cmd, usage_cmd, OutputFormat,
+};
+
+#[derive(Parser)]
+#[command(
+    name = "stella",
+    version = build_info::version_static(),
+    long_version = build_info::long_version_static(),
+    about = "A fast, BYOK, model-agnostic terminal coding agent"
+)]
+pub(crate) struct Cli {
+    #[command(flatten)]
+    pub(crate) globals: GlobalArgs,
+
+    #[command(subcommand)]
+    pub(crate) command: Option<Command>,
+}
+
+/// Session-wide flags shared by every subcommand — model routing,
+/// credentials, output shape, spend limit, UI toggles.
+///
+/// Every field here MUST carry `global = true`. clap accepts a plain
+/// root-level flag only *before* the subcommand token, so a non-global
+/// field in this struct is silently unreachable in the position users
+/// naturally type it (`stella fleet … --budget 5` dies with "unexpected
+/// argument"). `global = true` registers the flag with every subcommand,
+/// making both positions valid. The invariant is machine-enforced by
+/// `every_root_flag_is_global` in main_tests.rs — a new field without the
+/// attribute fails the suite, not a user's shell.
+///
+/// The names here are reserved CLI-wide: a subcommand flag reusing one
+/// does not shadow cleanly — clap propagates the global's value slot into
+/// every subcommand, and the id collision panics at match time in debug
+/// builds and misbinds in release. `no_subcommand_flag_reuses_a_global_name`
+/// in main_tests.rs enforces uniqueness (it is why `connect linear` pastes
+/// a key via `--paste-key`, not `--api-key`).
+#[derive(clap::Args)]
+pub(crate) struct GlobalArgs {
+    /// Override the worker model for this invocation: provider/model_id
+    /// (e.g. zai/glm-5.2, anthropic/claude-fable-5, openai/gpt-5.5)
+    #[arg(long, global = true, env = "STELLA_MODEL")]
+    pub(crate) model: Option<String>,
+
+    /// API key for the selected provider, highest-precedence step of the
+    /// credential chain (CLI flag -> env var -> credentials file ->
+    /// interactive prompt). Prefer an env var or
+    /// ~/.stella/credentials.toml for anything long-lived — a flag
+    /// value is visible in shell history and `ps`.
+    #[arg(long, global = true)]
+    pub(crate) api_key: Option<String>,
+
+    /// Base URL override. Required with `--model local/<model>` to point at a
+    /// local OpenAI-compatible server (Ollama, vLLM, LM Studio, llama.cpp
+    /// server — e.g. http://localhost:11434/v1); optional for every other
+    /// provider to route through a proxy.
+    #[arg(long, global = true, env = "STELLA_BASE_URL")]
+    pub(crate) base_url: Option<String>,
+
+    /// Output format: text (interactive), json (one final object), or
+    /// stream-json (one line per agent event)
+    #[arg(
+        long,
+        global = true,
+        env = "STELLA_OUTPUT_FORMAT",
+        value_enum,
+        default_value = "text"
+    )]
+    pub(crate) output_format: OutputFormat,
+
+    /// Hard USD spend limit for the whole run — a session-scoped cap:
+    /// enforced mode aborts cleanly (never mid-tool) once cumulative spend
+    /// across every turn and goal round exceeds this, not each turn on its
+    /// own. Omit to meter spend for the cost summary without ever blocking
+    /// (observed mode).
+    #[arg(long, global = true, env = "STELLA_BUDGET", value_parser = parse_budget)]
+    pub(crate) budget: Option<f64>,
+
+    /// Use the plain line-based REPL for chat instead of the Command Deck
+    /// (the tabbed TUI). The deck also steps aside automatically when stdin
+    /// or stdout is not a terminal. Env: STELLA_PLAIN=1.
+    #[arg(long, global = true)]
+    pub(crate) plain: bool,
+
+    /// Freeze all deck animation (the run progress bar's shimmer/pulse and the
+    /// caret blink) to a static frame — for CI and asciinema-style recordings.
+    /// Also forced on by STELLA_NO_ANIM or NO_COLOR.
+    #[arg(long, global = true)]
+    pub(crate) no_anim: bool,
+}
+
+#[derive(Subcommand)]
+pub(crate) enum Command {
+    /// Send a one-shot prompt (non-interactive)
+    Run {
+        /// The prompt to send
+        prompt: String,
+
+        /// Use the raw step-loop instead of the staged pipeline (triage, plan,
+        /// execute, verify, judge). The pipeline is the default; this flag
+        /// falls back to the direct Engine::run_turn path.
+        #[arg(long)]
+        no_pipeline: bool,
+
+        /// Test command the pipeline's verify stage runs deterministically
+        /// (e.g. "cargo test -p my-crate"). Arms the fail→pass flip oracle:
+        /// a change that flips a failing test to passing can submit without
+        /// a model-judge call. Omitted, verification always escalates to the
+        /// judge.
+        #[arg(long, value_name = "CMD")]
+        test_command: Option<String>,
+
+        /// Keep the authored witness test as a file in your working tree.
+        /// By default the witness is scaffolding: it proves this run's goal
+        /// inside the candidate workspace and is discarded with it, so an
+        /// already-satisfied test is never left behind in your test tree.
+        /// Pass this to promote it to a real test you can commit.
+        #[arg(long)]
+        keep_witness: bool,
+    },
+
+    /// arena-bench adapter: run the task in --task-dir (prompt in TASK.md)
+    /// while recording a contextgraph-trace journal the arena runner judges
+    /// with the protocol's replay oracles. Speaks the adapter contract
+    /// (--task-dir/--journal/--state-dir/--resume); see
+    /// <https://github.com/macanderson/arena-bench>.
+    Arena {
+        /// The episode workspace; the prompt is read from TASK.md inside it.
+        #[arg(long)]
+        task_dir: std::path::PathBuf,
+
+        /// The contextgraph-trace journal to append (crash-safe, per-event).
+        #[arg(long)]
+        journal: std::path::PathBuf,
+
+        /// Agent state that persists across episodes (memory arm).
+        #[arg(long)]
+        state_dir: std::path::PathBuf,
+
+        /// Present when re-invoked after a chaos kill: recover the journal,
+        /// declare what was recovered, continue the same session.
+        #[arg(long)]
+        resume: bool,
+
+        /// Use the raw step-loop instead of the staged pipeline.
+        #[arg(long)]
+        no_pipeline: bool,
+
+        /// Test command for the pipeline's deterministic verify ladder.
+        #[arg(long, value_name = "CMD")]
+        test_command: Option<String>,
+    },
+
+    /// Work in judged rounds until a judge model confirms the goal is met.
+    /// Each working round runs through the staged pipeline (triage, plan,
+    /// witness, execute, verify) by default; --no-pipeline falls back to the
+    /// raw step-loop.
+    Goal {
+        /// What must be true when done — assessed by the judge each round
+        goal: String,
+
+        /// Use the raw step-loop instead of the staged pipeline for each
+        /// working round. The pipeline is the default.
+        #[arg(long)]
+        no_pipeline: bool,
+    },
+
+    /// Watch CI for a branch/PR and fix failures until it is fully green
+    Monitor {
+        /// Branch name or PR number (default: main)
+        target: Option<String>,
+    },
+
+    /// Start an interactive REPL session
+    Chat,
+
+    /// Reopen a previous session exactly where it stood — transcript,
+    /// conversation, pending prompts. Sessions are durable (quit, crash, and
+    /// power loss included); the deck's SESSIONS overlay (`←` on an empty
+    /// prompt, `⏎` on a row) is the same navigation from inside a session.
+    Resume {
+        /// Registry id (`ses-…`) of the session to reopen. Omitted: the most
+        /// recently active resumable session of this workspace.
+        id: Option<String>,
+
+        /// List this machine's sessions (resumable ones marked) and exit.
+        #[arg(long)]
+        list: bool,
+    },
+
+    /// Analyze this workspace and infer its domain taxonomy
+    /// (.stella/domains.toml) — the tagging vocabulary for memories,
+    /// reflections, and every code-graph node/edge
+    Init,
+
+    /// List every tool available to the agent this session — built-ins,
+    /// developer custom tools (.stella/tools/), and manifest diagnostics
+    Tools {
+        /// Validate custom tool manifests instead of listing: parse every
+        /// `<name>.toml`, check names, required fields, timeouts, and
+        /// collisions with built-ins and other manifests, then exit
+        /// non-zero if any manifest has errors. Pass a directory to check
+        /// (defaults to the dirs discovery scans: .stella/tools/ and
+        /// ~/.stella/tools/).
+        #[arg(long, value_name = "DIR")]
+        validate: Option<Option<std::path::PathBuf>>,
+    },
+
+    /// Fan tasks out to a fleet of worker agents in ONE shared tree —
+    /// coordinated by cooperative claims (lock-on-first-write, sub-second,
+    /// rivals named), wave-scheduled by dependency, every attempt, commit,
+    /// and dollar recorded in .stella/private/fleet.db. Tasks opting into
+    /// isolation = "isolated" get a dedicated worktree whose `fleet/<task>`
+    /// branch is left in place for review.
+    Fleet {
+        /// Task prompts — each becomes an independent task in the SHARED
+        /// tree (cooperative claims coordinate writers; pass a plan file
+        /// with `isolation = "isolated"` for per-task worktrees)
+        #[arg(required_unless_present = "plan")]
+        tasks: Vec<String>,
+
+        /// A plan file instead: `.json` or `.toml` with `[[tasks]]` entries
+        /// (id, title, prompt, optional depends_on + isolation + claims —
+        /// paths held as cooperative file locks while the task runs)
+        #[arg(long, value_name = "FILE", conflicts_with = "tasks")]
+        plan: Option<std::path::PathBuf>,
+
+        /// Max tasks dispatched concurrently within one wave
+        #[arg(long, default_value_t = 4)]
+        max_concurrency: usize,
+
+        /// Git ref `isolation = "isolated"` worktrees branch from
+        /// (default: current HEAD); shared-tree tasks ignore it
+        #[arg(long)]
+        base_ref: Option<String>,
+
+        /// After the fan-out, watch each fleet branch's CI to completion and
+        /// reconcile its PR status via `gh` (the fleet PR/CI monitor). Exits
+        /// non-zero if any watched branch ends red. Meaningful once the
+        /// branches are pushed — e.g. task prompts that push and open PRs.
+        #[arg(long)]
+        watch: bool,
+
+        /// Use the raw step-loop instead of the staged pipeline (triage,
+        /// plan, witness, execute, verify) for each worker. The pipeline is
+        /// the default.
+        #[arg(long)]
+        no_pipeline: bool,
+    },
+
+    /// Query the code graph built by `stella init` — symbol definitions and
+    /// references, a file's imports/importers, or its graph neighborhood.
+    /// Offline: reads .stella/private/codegraph.db, needs no API key.
+    Graph {
+        /// What to ask the graph
+        #[arg(value_enum)]
+        op: contextgraph::GraphOp,
+
+        /// Symbol name (definitions/references) or workspace-relative file
+        /// path (imports/importers/neighbors)
+        target: String,
+    },
+
+    /// List or run the project's package-manager scripts — deterministic
+    /// static detection (cargo/npm/uv/go/make/just/…) mapped onto canonical
+    /// verbs (install/build/check/start/test/lint/format). Offline: manifest
+    /// parsing plus a local subprocess, needs no API key.
+    Scripts {
+        #[command(subcommand)]
+        cmd: scripts_cmd::ScriptsCmd,
+    },
+
+    /// Custom slash commands: list what this workspace offers, or convert
+    /// markdown definitions to TOML. Conversion is deliberate and never part
+    /// of `init` — `init` SYMLINKS `.claude/commands/`, so a converted copy
+    /// trades the live link for a typed `allowed-tools` and a delimiter-free
+    /// `prompt`. Offline: reads and writes local files, needs no API key.
+    Commands {
+        #[command(subcommand)]
+        cmd: commands_cmd::CommandsCmd,
+    },
+
+    /// Inspect the storage map — every storage layer, namespace, relation,
+    /// and field, with intent/boundaries from stella.storage.toml. Offline:
+    /// reads .stella/private/codegraph.db + the manifest, needs no API key.
+    Storage {
+        #[command(subcommand)]
+        cmd: storage_cmd::StorageCmd,
+    },
+
+    /// List configured providers and available models
+    Models {
+        #[command(subcommand)]
+        cmd: Option<ModelsCmd>,
+    },
+
+    /// Review what the adaptive-context loop wants to make durable, before it
+    /// lands: the proposals it induced, the distinct tasks behind each, and
+    /// Keep/Edit/Ignore. Every decision is recorded as an immutable event, so
+    /// the review history replays exactly. Offline: reads and appends to
+    /// .stella/private/context.db only, needs no API key. Phase 3 (#714).
+    Proposals {
+        #[command(subcommand)]
+        cmd: proposals_cmd::ProposalsCmd,
+    },
+
+    /// Review, publish, and explain context records — the reviewable steering
+    /// `stella ingest` extracts from documents you already wrote. `review` shows
+    /// what was proposed and what its truth probe found; `keep`/`edit`/`ignore`
+    /// decide; `list` shows what currently steers; `validate` re-probes every claim
+    /// and reports every finding; `explain` says why a rule applied; `propose` turns
+    /// a record into a reviewable change. Offline: local files only, no API key.
+    /// Epic #897.
+    Context {
+        #[command(subcommand)]
+        cmd: context_cmd::ContextCmd,
+    },
+
+    /// Eval-driven self-tuning of stella's own policy (#831): A/B one knob
+    /// (worker reasoning effort) over two loop-bench `--json` result files,
+    /// auto-select the winner, and — with `--promote` — write it to settings
+    /// with a reversible rollback record. `rollback` reverts the last
+    /// promotion; `status` shows the ledger. Offline: reads result files and
+    /// the local ledger, writes settings only on `--promote`. No API key.
+    Tune {
+        #[command(subcommand)]
+        cmd: tune_cmd::TuneCmd,
+    },
+
+    /// Show the exact context a past model call was sent — the reconstructed
+    /// message array, rebuilt from the recorded receipts and verified against
+    /// the digests taken at emission. With no arguments, lists executions that
+    /// have receipts; with an execution id, lists its model calls. Reads
+    /// .stella/private/store.db only; needs no API key and never writes.
+    Inspect {
+        /// Execution to inspect (omit to list recent executions)
+        execution_id: Option<i64>,
+
+        /// Turn instance within the execution
+        #[arg(long, default_value_t = 0)]
+        turn: u32,
+
+        /// Step to reconstruct (omit to list the execution's calls)
+        #[arg(long)]
+        step: Option<u64>,
+
+        /// Which model call at that step: 0 the worker, 1 the overflow
+        /// summarizer, 2+ a pipeline management role
+        #[arg(long = "call-seq", default_value_t = 0)]
+        call_seq: u64,
+
+        /// Output format
+        #[arg(long, value_enum, default_value = "text")]
+        format: inspect::InspectFormat,
+
+        /// Print message bodies in full instead of eliding long ones
+        #[arg(long)]
+        full: bool,
+
+        /// Show a unified diff of what changed instead of the whole context.
+        /// Bare `--diff` compares against whatever ran immediately before in
+        /// the same role — the previous step, or the previous turn when this
+        /// is a turn's first call. The session's very first call has no
+        /// predecessor, so it is compared against the prompt as the user
+        /// submitted it: the mutation no other view shows
+        #[arg(long, value_enum, num_args = 0..=1, default_missing_value = "prev")]
+        diff: Option<inspect::DiffBase>,
+
+        /// Unchanged lines to print around each change in a `--diff`
+        #[arg(long, default_value_t = 3)]
+        context: usize,
+
+        /// Restrict `--diff` to one message role — `system` is the usual
+        /// reason to reach for this, since every step appends tool traffic
+        /// that would otherwise bury the prompt's own delta
+        #[arg(long, value_enum, default_value = "all")]
+        only: inspect::RoleFilter,
+    },
+
+    /// Summarize cost, tokens, and resolve rate per provider/model from
+    /// local telemetry (.stella/private/store.db) — $/resolved-task receipts.
+    /// `stella stats prune` bounds that store's growth
+    Stats {
+        /// Output format: table (aligned, with TOTAL row), json, or csv
+        #[arg(long, value_enum, default_value = "table")]
+        format: stats::StatsFormat,
+
+        /// Only show executions for this provider id (e.g. zai, anthropic,
+        /// local)
+        #[arg(long)]
+        provider: Option<String>,
+
+        #[command(subcommand)]
+        cmd: Option<stats::StatsCmd>,
+    },
+
+    /// Cross-project telemetry hub (~/.stella/usage.db): global report,
+    /// cursor-based sync, backfill across every known project
+    Usage {
+        #[command(subcommand)]
+        cmd: Option<usage_cmd::UsageCmd>,
+    },
+
+    /// Cloud account registration (stub): org / workspace identity that
+    /// scopes replicated telemetry; OAuth login attaches here later
+    Cloud {
+        #[command(subcommand)]
+        cmd: usage_cmd::CloudCmd,
+    },
+
+    /// Inspect or explicitly flush the managed enterprise operational spool.
+    /// Disabled by default; requires a signed org-managed enrollment.
+    Telemetry {
+        #[command(subcommand)]
+        cmd: TelemetryCmd,
+    },
+
+    /// Open the Observatory — a local web dashboard over this workspace's
+    /// telemetry (.stella/private/store.db + fleet.db): spend, tokens, cache
+    /// traffic, tool calls, files touched, memory citations, reflections,
+    /// and fleet runs. Binds 127.0.0.1 only and opens the stores strictly
+    /// read-only — nothing ever leaves this machine.
+    Observe {
+        /// Port to bind on 127.0.0.1 (0 picks a free port)
+        #[arg(long, default_value_t = 7787)]
+        port: u16,
+
+        /// Open the dashboard in the default browser once serving
+        #[arg(long)]
+        open: bool,
+    },
+
+    /// Turn markdown you already wrote — `AGENTS.md`, `CLAUDE.md`, or any file
+    /// you name — into steering Stella can check, cite, and retire. With no
+    /// arguments, scans the workspace and shows what it found. Reads local
+    /// files only; needs no API key.
+    Ingest(ingest_cmd::IngestArgs),
+
+    /// What the work cost, and whether anyone said it was good: calls, characters
+    /// typed, follow-ups, and the verdict a merged or closed pull request implies.
+    /// Reads local state only; needs no API key, and no model judges anything.
+    Scoreboard,
+
+    /// Inspect the project's memories through the citation feedback loop —
+    /// most-cited first, usefulness scores, truthfulness — and promote an
+    /// eligible memory to a project rule (.stella/rules/). Reads local state
+    /// only; needs no API key.
+    Memory {
+        #[command(subcommand)]
+        cmd: memory_cmd::MemoryCmd,
+    },
+
+    /// Manage MCP servers: search a registry, install into .stella/mcp.toml,
+    /// list configured servers, and show tool-usage telemetry. Enable/disable
+    /// is per-session and lives in the deck's MCP tab (`/mcp`). Reads/writes
+    /// local state (+ the registry over HTTP); needs no API key.
+    Mcp {
+        #[command(subcommand)]
+        cmd: McpCmd,
+    },
+
+    /// Connect an issue tracker (GitHub/Linear) via OAuth or a pasted key —
+    /// enables the issue tools (search_issues, create_issue, list_labels, …)
+    /// and the deck's Issues tab. Credentials land owner-only in
+    /// ~/.stella/integrations.json; needs no model API key.
+    Connect {
+        #[command(subcommand)]
+        cmd: ConnectCmd,
+    },
+
+    /// Show current configuration
+    Config,
+
+    /// Check the local state stella owns and report each named check's
+    /// verdict — today the integrity of this workspace's session store
+    /// (.stella/private/store.db), verified with SQLite's own
+    /// quick_check/integrity_check. Exits non-zero if any check fails, so it
+    /// can gate a script. Reads local state only; needs no API key.
+    Doctor {
+        /// Repair a store.db that failed the check: move it aside to a
+        /// timestamped name (RENAMED, never deleted — its WAL/SHM siblings
+        /// travel with it), then copy out whatever is still readable into a
+        /// separate salvaged database. Only ever acts on a database SQLite
+        /// itself judged corrupt; a healthy store and an inconclusive check
+        /// are both left untouched. The next session starts a fresh store.
+        #[arg(long)]
+        repair: bool,
+    },
+
+    /// Manage BYOK provider keys stored in
+    /// ~/.stella/credentials.toml (set/remove/list) — keys resolved
+    /// via an env var or settings.json still take precedence per the normal
+    /// chain; `stella models`/`stella config` show which source actually
+    /// wins. Never prints a secret value; needs no model API key itself.
+    Auth {
+        #[command(subcommand)]
+        cmd: AuthCmd,
+    },
+
+    /// Print a shell completion script for stella to stdout.
+    ///
+    /// Install it the way your shell expects, e.g.
+    ///
+    ///     bash: stella completions bash > /etc/bash_completion.d/stella
+    ///
+    ///     zsh:  stella completions zsh  > "${fpath[1]}/_stella"
+    ///
+    ///     fish: stella completions fish > ~/.config/fish/completions/stella.fish
+    ///
+    /// Offline, writes nothing, and needs no API key.
+    Completions {
+        /// bash | zsh | fish | powershell | elvish
+        #[arg(value_enum)]
+        shell: clap_complete::Shell,
+    },
+
+    /// Print the version and exit
+    Version,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Subcommand)]
+pub(crate) enum TelemetryCmd {
+    /// Show enrollment state, pending bytes/rows, and local drop count.
+    Status,
+    /// Attempt one bounded delivery batch now.
+    Flush,
+    /// Explicitly discard rows bound to a superseded enrolled sink.
+    RolloverDiscard,
+}
+
+/// `stella auth` subcommands — the whole `~/.stella/credentials.toml`
+/// management surface. Deliberately small: a handful of BYOK keys, not a
+/// config language (mirrors `CredentialsFile`'s own doc intent).
+#[derive(Subcommand)]
+pub enum AuthCmd {
+    /// Store (or replace) a provider's API key in credentials.toml.
+    Set {
+        /// Provider id (a built-in like `zai`/`anthropic`/`openai`, or a
+        /// settings.json-defined custom provider id)
+        provider: String,
+
+        /// Pass the key directly on the command line. WARNING: this value
+        /// becomes visible in shell history and `ps` output — prefer
+        /// --stdin, or omit both flags for an interactive masked prompt.
+        #[arg(long, conflicts_with = "stdin")]
+        key: Option<String>,
+
+        /// Read the key from stdin (one line, trimmed) instead of a flag or
+        /// an interactive prompt — for scripts, e.g. `printf '%s' "$KEY" | \
+        /// stella auth set zai --stdin`.
+        #[arg(long)]
+        stdin: bool,
+    },
+
+    /// Remove a provider's stored key from credentials.toml.
+    Remove {
+        /// Provider id
+        provider: String,
+    },
+
+    /// List providers with a key stored in credentials.toml (redacted
+    /// preview + resolution source).
+    List,
+}
+
+/// `stella models` subcommands — the model-catalog surface. A bare
+/// `stella models` keeps its provider/key listing; these manage the
+/// on-disk master list every slug validates against.
+#[derive(Subcommand)]
+pub(crate) enum ModelsCmd {
+    /// Sync the model catalog: the models.dev master list (public, no API
+    /// key), then every configured provider's own live /models listing.
+    /// Incremental: an unchanged master list is one conditional request
+    /// (ETag) and zero writes; pricing changes append a new model-card
+    /// version (the latest version is what displays everywhere). Also runs
+    /// automatically once a day while a provider credential is configured.
+    Refresh {
+        /// Re-download even when the server says the list is unchanged
+        /// (recovery hatch for a corrupted local catalog)
+        #[arg(long)]
+        force: bool,
+    },
+    /// List the model catalog: provider/model slugs with latest pricing
+    /// (USD per Mtok), context window, capability, and model maker. Scoped
+    /// to providers whose credential resolves; --all lifts the scope.
+    List {
+        /// Only this provider id (e.g. anthropic, openrouter)
+        #[arg(long)]
+        provider: Option<String>,
+
+        /// Include providers with no configured credential
+        #[arg(long)]
+        all: bool,
+    },
+}
+
+/// `stella connect` subcommands — tracker connections consumed by the issue
+/// tools. GitHub uses the OAuth device flow (public client, no secret in the
+/// binary); Linear uses browser OAuth when an app is configured, else a
+/// personal API key. All traffic is user-initiated — connecting is what opts
+/// a workspace into tracker calls.
+#[derive(Subcommand)]
+pub enum ConnectCmd {
+    /// Connect GitHub via the OAuth device flow (or --token to paste a PAT)
+    Github {
+        /// Paste a personal access token instead of running the device flow
+        #[arg(long)]
+        token: bool,
+    },
+    /// Connect Linear via browser OAuth (needs STELLA_LINEAR_CLIENT_ID) or a
+    /// personal API key
+    Linear {
+        /// Paste a personal API key even when an OAuth app is configured.
+        /// (Named to stay clear of the session-wide `--api-key`, which is
+        /// the model-provider credential — a different secret entirely.)
+        #[arg(long)]
+        paste_key: bool,
+    },
+    /// Show stored connections, their accounts, and credential precedence
+    Status,
+    /// Forget a stored connection
+    Remove {
+        /// github | linear
+        provider: String,
+    },
+}
+
+/// `stella mcp` subcommands — the scriptable half of the MCP management surface
+/// (the deck's MCP tab is the interactive half; per-session enable/disable and
+/// the masked auth prompt live only there).
+#[derive(Subcommand)]
+pub enum McpCmd {
+    /// List configured MCP servers (.stella/mcp.toml)
+    List,
+    /// Search the MCP server registry (settings.json `mcp.registry_url`, else
+    /// the official registry)
+    Search {
+        /// Substring to match server names (omit to list)
+        query: Vec<String>,
+        /// Max results in the page
+        #[arg(long)]
+        limit: Option<u32>,
+    },
+    /// Install a registry server into .stella/mcp.toml (overwrites — MCP
+    /// servers are not versioned)
+    Install {
+        /// The registry server name (as shown by `stella mcp search`)
+        name: String,
+        /// Local alias / tool-namespace segment (default: sanitized name)
+        #[arg(long)]
+        alias: Option<String>,
+    },
+    /// Remove a configured server from .stella/mcp.toml
+    Remove {
+        /// The configured server's local name
+        name: String,
+    },
+    /// OAuth login to a configured http server (opens your browser; tokens
+    /// land owner-only in .stella/private/mcp_oauth.json and auto-refresh)
+    Login {
+        /// The configured server's local name
+        name: String,
+    },
+    /// Forget a server's OAuth tokens
+    Logout {
+        /// The configured server's local name
+        name: String,
+    },
+    /// Show MCP tool-usage telemetry (.stella/private/store.db): calls per server/tool
+    Usage,
+}
+
+/// `--budget` must be a positive, finite dollar amount — a NaN or negative
+/// limit would make every comparison silently false and turn the "hard
+/// cap" into a no-op, the worst failure mode for a money control.
+fn parse_budget(raw: &str) -> Result<f64, String> {
+    let value: f64 = raw
+        .parse()
+        .map_err(|_| format!("`{raw}` is not a number"))?;
+    if !value.is_finite() || value <= 0.0 {
+        return Err(format!(
+            "budget must be a positive dollar amount, got `{raw}`"
+        ));
+    }
+    Ok(value)
+}
